SW-4740: spi: change mode type to uint16_t
The mode parameters used in Greybus transfers, struct
device_spi_device_config, struct device_spi_master_config, and struct
spi_board_info are all uint16_t. Change the function prototype in
struct device_spi_type_ops and the implementation in tsb_spi.c to use a
uint16_t mode.

Testing Done: compiled

+++ nuttx/arch/arm/src/tsb/tsb_spi.c
@@ -482,7 +482,7 @@ err_freq_set:
  * @param mode SPI protocol mode requested
  * @return 0 on success, negative errno on error
  */
-static int tsb_spi_setmode(struct device *dev, uint8_t devid, uint8_t mode)
+static int tsb_spi_setmode(struct device *dev, uint8_t devid, uint16_t mode)
 {
     struct tsb_spi_dev_info *info = NULL;
     uint32_t ctrl0 = 0;
